check if path exists when local repository paths are used
The test for existance of path was combined with the test for the
presence of URI pattern(://) to check if the repository was local or
remote. Even if one of them failed, it would be considered as a remote
repository and nodegit threw an error (unsupported URL protocol).

Now, the checks have been separated and an error is thrown to indicate
that the path does not exist.